include tests for initial document trace-id
This includes tests highlighting the issue described by
https://datadoghq.atlassian.net/browse/RUMF-670
